backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
ifcopenshell
openpyxl==3.1.5
openai==1.50.0
python-dotenv==1.0.1
